requirements.txt
absl-py==1.0.0
blobfile==2.0.2
clean-fid==0.1.35
einops==0.6.0
glob2==0.7
kornia==0.6.12
matplotlib==3.5.0
numpy==1.24.2
pandas==1.3.4
pathlib==1.0.1
Pillow==9.5.0
scikit-image==0.20.0
cikit-learn==0.24.0
scipy==1.9.1
statistics
tensorboard==2.7.0
tensorly==0.8.1
tensorly-torch==0.4.0
torch==1.13.1
torchgeometry==0.1.2
torchvision==0.14.1
tqdm==2.2.3
transformers==4.27.4
